Updated according to PR requests
