+++ source/frontend/src/utils/personUtils.ts
@@ -1,4 +1,3 @@
-import { IContactSearchResult } from '@/interfaces';
 import { IEditablePerson } from '@/interfaces/editable-contact';
 import { Api_Person } from '@/models/api/Person';
 
@@ -15,17 +14,4 @@ export function formatApiPersonNames(person?: Api_Person | null): string {
 
 export function formatNames(nameParts: Array<string | undefined | null>): string {
   return nameParts.filter(n => n !== null && n !== undefined && n.trim() !== '').join(' ');
-}
-
-export function formatContactSearchResult(
-  contact: IContactSearchResult,
-  defaultText: string = '',
-): string {
-  let text = defaultText;
-  if (contact?.personId !== undefined) {
-    text = formatNames([contact.firstName, contact.middleNames, contact.surname]);
-  } else if (contact?.organizationId !== undefined) {
-    text = contact.organizationName || '';
-  }
-  return text;
 }
